backend/requirements.txt
alembic==1.9.2
auth==0.5.3
blinker==1.5
certifi==2022.12.7
charset-normalizer==3.0.1
click==7.1.2
dnspython==2.3.0
ecdsa==0.18.0
eventlet==0.33.3
falcon==3.1.1
fastapi==0.68.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
Mako==1.2.4
MarkupSafe==2.1.2
mongoengine==0.26.0
passlib==1.7.4
psycopg2-binary==2.9.5
pyasn1==0.4.8
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-jose==3.3.0
requests==2.28.2
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.1
starlette==0.14.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.12.2
Werkzeug==2.2.2
zipp==3.12.0
